feat(setup): webpack setup fro client side code
mode sets process.env.NODE_ENV to the given value and tells Webpack
to use its built-in optimizations accordingly. If not set explicitly, it defaults
to a value of "production". It can also be set via the command line by
passing the value as a CLI argument.
devtool specifies how source maps are generated, if at all. Generally, a
source map provides a way of mapping code within a compressed file back
to its original position in a source file to aid debugging.
entry specifies the entry file where Webpack starts bundling, in this case
with the main.js file in the client folder.
output specifies the output path for the bundled code, in this case set to
dist/bundle.js.
publicPath allows specifying the base path for all assets in the
application.
module sets the regex rule for the file extension to be used for transpilation,
and the folders to be excluded. The transpilation tool to be used here is
babel-loader.
HotModuleReplacementPlugin enables hot module replacement for
react-hot-loader.
NoEmitOnErrorsPlugin allows skipping emitting when there are compile
errors.
We also add a Webpack alias to point react-dom references to the @hot-loader/react-dom version