feat(setup): webpack server side
The mode option is not set here explicitly but will be passed as required when
running the Webpack commands with respect to running for development or
building for production.
Webpack starts bundling from the server folder with server.js, then outputs the
bundled code in server.generated.js in the dist folder. During bundling, a
CommonJS environment will be assumed as we are specifying commonjs2
in libraryTarget, so the output will be assigned to module.exports